+++ packages/web/src/components/range/DynamicRangeSlider.js
@@ -77,7 +77,13 @@ class DynamicRangeSlider extends Component {
 					Math.ceil(nextProps.range.end),
 				], nextProps);
 			}
-		} else if (!isEqual(this.props.defaultSelected, nextProps.defaultSelected) && nextProps.range) {
+		} else if (nextProps.range
+			&& !isEqual(
+				this.props.defaultSelected
+				&& this.props.defaultSelected(nextProps.range.start, nextProps.range.end),
+				nextProps.defaultSelected
+				&& nextProps.defaultSelected(nextProps.range.start, nextProps.range.end),
+			)) {
 			const { start, end } = nextProps.defaultSelected(nextProps.range.start, nextProps.range.end);
 			this.handleChange(
 				[start, end],